+++ src/imp/interrupt/mod.rs
@@ -41,7 +41,10 @@ mod arch;
 
 use core::{cell::UnsafeCell, sync::atomic::Ordering};
 
-const IS_ALWAYS_LOCK_FREE: bool = false;
+// Consider atomic operations based on disabling interrupts on single-core
+// systems are lock-free. (We consider the pre-v6 ARM Linux's atomic operations
+// provided in a similar way by the Linux kernel to be lock-free.)
+const IS_ALWAYS_LOCK_FREE: bool = true;
 
 #[inline]
 fn with<F, R>(f: F) -> R